sales/requirements.txt
fastapi==0.120.0
uvicorn[standard]==0.38.0
pdfplumber==0.11.7
gspread==6.2.1
google-auth==2.41.1
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.150.0
python-multipart==0.0.20
jinja2==3.1.6
python-dotenv==1.1.1
pydantic>=2.8.0
sqlalchemy>=2.0.30
psycopg2-binary
alembic>=1.16.0
pillow>=10.0.0
pytesseract==0.3.10
numpy>=1.26.0
opencv-python-headless>=4.9.0.80
httpx==0.27.0
itsdangerous==2.2.0
pillow-heif==0.16.0
requests==2.32.5
pytz==2024.2
rapidocr-onnxruntime>=1.3.7
